Rename gr to groot; make it actually work.
In the previous version, the .zshrc took $(git top) only at the time
of sourcing, which made it seem like it was behaving with one test, but
beyond that would constantly point to the same location.

Using a function as an alternative allows this to be evaluated each time.

Also, functions as a directory and installation weren't there, so that was
added.

Further, this had to be renamed from "gr" to "groot" as the oh-my-zsh git
plugin oddly defines a "gr" and other extended variants that I didn't want
to mess with removing at this time. Frankly, "groot" is just a better, more
memorable name anyway. I am, after all, groot.